requirements.txt
gradio
openai
python-dotenv
sentence-transformers
faiss-cpu
pandas
pdfminer.six
requests
numpy
